2026-05-30 14:20 CST The issues with cfsecram have been resolved, and with that clean compiling, the rest of 3.1 clean compiled.  The code is finally back into a testable, runnable state.

+++ src/main/java/server/markhome/mcf/v3_1/cfsec/cfsecram/CFSecRamISOCcyTable.java
@@ -330,7 +330,7 @@ public class CFSecRamISOCcyTable
 		ICFSecISOCcy iBuff )
 	{
 		CFSecBuffISOCcy Buff = (CFSecBuffISOCcy)ensureRec(iBuff);
-		Short pkey = Buff.getPKey();
+		Short pkey = (Short)Buff.getPKey();
 		CFSecBuffISOCcy existing = dictByPKey.get( pkey );
 		if( existing == null ) {
 			throw new CFLibStaleCacheDetectedException( getClass(),
