+++ src/main/java/server/markhome/mcf/v3_1/cfsec/cfsecram/CFSecRamISOLangTable.java
@@ -345,7 +345,7 @@ public class CFSecRamISOLangTable
 		ICFSecISOLang iBuff )
 	{
 		CFSecBuffISOLang Buff = (CFSecBuffISOLang)ensureRec(iBuff);
-		Short pkey = Buff.getPKey();
+		Short pkey = (Short)Buff.getPKey();
 		CFSecBuffISOLang existing = dictByPKey.get( pkey );
 		if( existing == null ) {
 			throw new CFLibStaleCacheDetectedException( getClass(),
